+++ mycoffee/functions.py
@@ -4,7 +4,7 @@ import math
 from mycoffee.params import MESSAGE_TEMPLATE, METHODS_LIST_TEMPLATE, EMPTY_INFO
 from mycoffee.params import MY_COFFEE_VERSION, DEFAULT_PARAMS
 from mycoffee.params import METHODS_MAP, COFFEE_UNITS_MAP, WATER_UNITS_MAP
-from mycoffee.params import RATIO_WARNING_MESSAGE
+from mycoffee.params import RATIO_WARNING_MESSAGE, GRIND_WARNING_MESSAGE
 from art import tprint
 
 
@@ -47,6 +47,10 @@ def print_result(params):
         ratio_lower_limit = METHODS_MAP[method]["ratio_lower_limit"]
         ratio_upper_limit = METHODS_MAP[method]["ratio_upper_limit"]
         print(RATIO_WARNING_MESSAGE.format(method, str(ratio_lower_limit), str(ratio_upper_limit)))
+    if not check_grind_limits(params):
+        grind_lower_limit = METHODS_MAP[method]["grind_lower_limit"]
+        grind_upper_limit = METHODS_MAP[method]["grind_upper_limit"]
+        print(GRIND_WARNING_MESSAGE.format(method, str(grind_lower_limit), str(grind_upper_limit)))
 
 
 def load_method_params(method_name):
@@ -167,6 +171,24 @@ def check_ratio_limits(params):
         ratio_lower_limit = METHODS_MAP[method]["ratio_lower_limit"]
         ratio_upper_limit = METHODS_MAP[method]["ratio_upper_limit"]
         if ratio < ratio_lower_limit or ratio > ratio_upper_limit:
+            return False
+    return True
+
+
+def check_grind_limits(params):
+    """
+    Check grind limits.
+
+    :param params: parameters
+    :type params: dict
+    :return: result as bool (False --> the grind is out of range)
+    """
+    method = params["method"]
+    if "grind_lower_limit" in METHODS_MAP[method] and "grind_upper_limit" in METHODS_MAP[method]:
+        grind = params["grind"]
+        grind_lower_limit = METHODS_MAP[method]["grind_lower_limit"]
+        grind_upper_limit = METHODS_MAP[method]["grind_upper_limit"]
+        if grind < grind_lower_limit or grind > grind_upper_limit:
             return False
     return True