+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.2] - 2025-01-13
 ### Added
 - 2 new water units
 	1. Troy Pounds (`t lb`)
@@ -140,7 +141,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 	5. Siphon
 	6. Custom
 
-[Unreleased]: https://github.com/sepandhaghighi/mycoffee/compare/v1.1...dev
+[Unreleased]: https://github.com/sepandhaghighi/mycoffee/compare/v1.2...dev
+[1.2]: https://github.com/sepandhaghighi/mycoffee/compare/v1.1...v1.2
 [1.1]: https://github.com/sepandhaghighi/mycoffee/compare/v1.0...v1.1
 [1.0]: https://github.com/sepandhaghighi/mycoffee/compare/v0.9...v1.0
 [0.9]: https://github.com/sepandhaghighi/mycoffee/compare/v0.8...v0.9